data/requirements.txt
Pillow
pytesseract
opencv-python
deskew
